ParliamentaryDiarizer/requirements.txt
opencv-python~=4.8.1.78
torch>=2.1.0
torchvision~=0.16.0
matplotlib~=3.5.1
numpy~=1.24.4
scikit-learn~=1.0.2
Pillow~=9.0.1
pandas~=1.4.1
sklearn~=0.0
setuptools~=57.4.0
tensorflow~=2.11.0
tqdm~=4.64.0
h5py~=3.8.0
transformers~=4.39.3
pyannote.audio==3.1.1
openai-whisper==20231117
openai==1.30.2
pyqt5==5.15.9